+++ src/main/index.ts
@@ -1,5 +1,6 @@
-export * from "./classes";
 export * from "./activity";
+export * from "./client";
+export * from "./loader";
 export * from "./readline";
 export * from "./styles";
 export * from "./time";