+++ tests/index.spec.ts
@@ -1,8 +1,9 @@
 import sinon from "sinon";
 sinon.stub(console, "log");
 
-import "./loader.spec";
 import "./activity.spec";
 import "./command.spec";
-import "./template.spec";
+import "./loader.spec";
+import "./store.spec";
 import "./styles.spec";
+import "./template.spec";