+++ CHANGELOG.md
@@ -6,10 +6,18 @@ All notables change for Cocoa Discord Utils will be noted here
 
 **Note 2**: Not all change will be noted, see commits for *full* changelog
 
-## [1.0.0] - 2022-02-2?
+## [1.0.0] - 2022-02-22
 
 - @MessageCommand (CogMessageClass) no longer require 'name' field.
 (You can still specify it to override what ever your method name is)
+
+- Added `checkLogin()`
+
+- Added ConsoleManager for more simple console command handling
+
+- Added `setReloadInterval` method for Loader
+
+- Miscellanous Changes
 
 ## [1.0.0-rc.5] - 2022-02-21